Prefer Numerics in schema config over Floats
* 'Float' will throw errors if an Integer is given to the config

Authored-by: Seth Boyles <sboyles@pivotal.io>

+++ lib/cloud_controller/config_schemas/base/api_schema.rb
@@ -184,7 +184,7 @@ module VCAP::CloudController
             broker_client_timeout_seconds: Integer,
             broker_client_default_async_poll_interval_seconds: Integer,
             broker_client_max_async_poll_duration_minutes: Integer,
-            broker_client_async_poll_exponential_backoff_rate: Float,
+            broker_client_async_poll_exponential_backoff_rate: Numeric,
             optional(:uaa_client_name) => String,
             optional(:uaa_client_secret) => String,
             optional(:uaa_client_scope) => String,
